django/requirements.txt
Django==5.1.1
django-ninja-extra==0.21.8
django-cors-headers==4.6.0
psycopg==3.2.3